requirements.txt
pandas
openpyxl
yagmail
python-dotenv
